# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Portuguese Autónomo & Tax Optimization API",
    "version": "2.0.0",
    "description": "Local-first operations & accounting assistant for Portuguese Autónomos. Two layers: (1) stateless CIRS/CRCSPSS/CIVA tax & Social Security simulators — Recibos Verdes, e-Fatura deductions, Segurança Social quarterly variation (Art. 163º CRCSPSS), and Unipessoal Lda break-even comparison; (2) a persistent profile + transaction ledger that auto-derives a personalized compliance obligations timeline and dashboard, so numbers don't have to be re-entered by hand on every call. This is a local system of record and calculator — it does not file anything with the AT or Segurança Social on your behalf.",
    "contact": {
      "name": "Portuguese Autónomos Open Source Community",
      "url": "https://github.com/n4ouri/pt-autonomo-api"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/api/v1/profiles": {
      "post": {
        "summary": "Create an autónomo profile",
        "description": "Creates the persistent profile everything else (ledger, obligations, dashboard) is scoped to. Returns the API key exactly once — store it, it cannot be recovered.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Maria Silva"
                  },
                  "nif": {
                    "type": "string",
                    "example": "123456789"
                  },
                  "activityStartDate": {
                    "type": "string",
                    "format": "date",
                    "example": "2026-01-01"
                  },
                  "regimeIVA": {
                    "type": "string",
                    "enum": [
                      "isento",
                      "trimestral",
                      "mensal"
                    ],
                    "example": "trimestral"
                  },
                  "serviceType": {
                    "type": "string",
                    "enum": [
                      "SERVICOS_PROFISSIONAIS",
                      "OUTROS_SERVICOS",
                      "VENDAS_MERCADORIAS"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Profile created; apiKey included once"
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "summary": "Get the authenticated profile",
        "description": "Returns the profile identified by the bearer API key.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "summary": "Update profile settings",
        "description": "Partially updates name/nif/activityStartDate/regimeIVA/serviceType.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "nif": {
                    "type": "string"
                  },
                  "activityStartDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "regimeIVA": {
                    "type": "string",
                    "enum": [
                      "isento",
                      "trimestral",
                      "mensal"
                    ]
                  },
                  "serviceType": {
                    "type": "string",
                    "enum": [
                      "SERVICOS_PROFISSIONAIS",
                      "OUTROS_SERVICOS",
                      "VENDAS_MERCADORIAS"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/me/transactions": {
      "get": {
        "summary": "List transactions",
        "description": "Lists the profile's income/expense ledger, filterable by from/to/type/category query params.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "summary": "Log a transaction",
        "description": "Adds an income or expense entry. Income category drives the CIRS coefficient and Social Security calculation.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "type",
                  "date",
                  "amount",
                  "category"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "income",
                      "expense"
                    ]
                  },
                  "date": {
                    "type": "string",
                    "format": "date",
                    "example": "2026-02-10"
                  },
                  "amount": {
                    "type": "number",
                    "example": 1500
                  },
                  "category": {
                    "type": "string",
                    "example": "SERVICOS_PROFISSIONAIS"
                  },
                  "vatRate": {
                    "type": "number",
                    "example": 0.23
                  },
                  "counterpartyNIF": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "isJustifiedExpense": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction created"
          }
        }
      }
    },
    "/api/v1/me/transactions/{id}": {
      "patch": {
        "summary": "Update a transaction",
        "description": "Partially updates a transaction owned by the authenticated profile.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "summary": "Delete a transaction",
        "description": "Deletes a transaction owned by the authenticated profile.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        }
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
    "/api/v1/me/obligations": {
      "get": {
        "summary": "List personalized obligations",
        "description": "Derives the 'what's due, when, how much' timeline from the statutory TAX_CALENDAR, the profile's regime, and its ledger, overlaid with any stored paid/pending status.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "upcomingDays",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Only return obligations due within this many days."
          }
        ]
      }
    },
    "/api/v1/me/obligations/{key}": {
      "patch": {
        "summary": "Update an obligation's status",
        "description": "Marks a derived obligation as paid/pending and records amountPaid/paidDate/notes.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "pending",
                      "paid"
                    ]
                  },
                  "amountPaid": {
                    "type": "number"
                  },
                  "paidDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "parameters": [
        {
          "name": "key",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key returned once by POST /api/v1/profiles, sent as 'Authorization: Bearer <apiKey>'."
      }
    }
  }
}